+++ waspc/cli/src/Wasp/Cli/Command/CreateNewProject/StarterTemplates/GhReleaseArchive.hs
@@ -18,6 +18,8 @@ import Wasp.Cli.GithubRepo
     fetchFolderFromGithubReleaseArchiveToDisk,
   )
 import Wasp.Project (WaspProjectDir)
+import Wasp.Util (indent)
+import Wasp.Util.InstallMethod (getInstallationCommand)
 
 createProjectOnDiskFromGhReleaseArchiveTemplate ::
   String ->
@@ -52,7 +54,7 @@ createProjectOnDiskFromGhReleaseArchiveTemplate templateName absWaspProjectDir p
             "Visit " ++ releasesUrl ++ " to see available template releases,",
             "and install the Wasp version that matches the latest release tag available by running:",
             "",
-            "  curl -sSL https://get.wasp.sh/installer.sh | sh -s -- -v x.y.z",
+            indent 2 $ getInstallationCommand (Just "x.y.z"),
             "",
             "Then you can try creating your project again."
           ]