+++ waspc/data/Cli/starters/basic/src/App.tsx
@@ -1,4 +1,4 @@
-import { Outlet } from "react-router-dom";
+import { Outlet } from "react-router";
 import "./App.css";
 import { Header } from "./shared/components/Header";