+++ waspc/e2e-tests/snapshots/kitchen-sink-golden/wasp-app/.wasp/out/web-app/src/auth/pages/createAuthRequiredPage.jsx
@@ -1,6 +1,6 @@
 import React from 'react'
 
-import { Navigate } from 'react-router-dom'
+import { Navigate } from 'react-router'
 import { useAuth } from 'wasp/client/auth'
 
 import { Loader } from '../../components/Loader'